refactor: split sparse vector index
