+++ lib/segment/src/index/sparse_index/sparse_vector_index.rs
@@ -1,45 +1,31 @@
 use std::borrow::Cow;
-use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 
 use atomic_refcell::AtomicRefCell;
+#[cfg(feature = "testing")]
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::generic_consts::Random;
-use common::storage_version::{StorageVersion as _, VERSION_FILE};
-use common::types::{PointOffsetType, ScoredPointOffset, TelemetryDetail};
+use common::storage_version::StorageVersion as _;
+use common::types::PointOffsetType;
 use fs_err as fs;
-use itertools::Itertools;
 use semver::Version;
 use sparse::common::scores_memory_pool::ScoresMemoryPool;
 use sparse::common::sparse_vector::SparseVector;
-use sparse::common::types::DimId;
 use sparse::index::inverted_index::inverted_index_ram_builder::InvertedIndexBuilder;
 use sparse::index::inverted_index::{INDEX_FILE_NAME, InvertedIndex, OLD_INDEX_FILE_NAME};
-use sparse::index::search_context::SearchContext;
 
 use super::indices_tracker::IndicesTracker;
-use super::sparse_index_config::SparseIndexType;
 use crate::common::operation_error::{OperationError, OperationResult, check_process_stopped};
-use crate::common::operation_time_statistics::ScopeDurationMeasurer;
-use crate::data_types::named_vectors::CowVector;
-use crate::data_types::query_context::VectorQueryContext;
-use crate::data_types::vectors::{QueryVector, VectorInternal, VectorRef};
 use crate::id_tracker::{IdTrackerEnum, IdTrackerRead};
-use crate::index::field_index::CardinalityEstimation;
-use crate::index::hnsw_index::point_scorer::BatchFilteredSearcher;
-use crate::index::query_estimator::adjust_to_available_vectors;
 use crate::index::sparse_index::sparse_index_config::SparseIndexConfig;
 use crate::index::sparse_index::sparse_search_telemetry::SparseSearchesTelemetry;
 use crate::index::struct_payload_index::StructPayloadIndex;
-use crate::index::{PayloadIndexRead, VectorIndex, VectorIndexRead};
-use crate::telemetry::VectorIndexSearchesTelemetry;
-use crate::types::{DEFAULT_SPARSE_FULL_SCAN_THRESHOLD, Filter, SearchParams};
-use crate::vector_storage::query::TransformInto;
-use crate::vector_storage::{
-    VectorStorage, VectorStorageEnum, VectorStorageRead, check_deleted_condition,
-};
+use crate::vector_storage::{VectorStorageEnum, VectorStorageRead};
+
+mod search;
+mod vector_index_impl;
 
 /// Whether to use the new compressed format.
 pub const USE_COMPRESSED: bool = true;
@@ -273,460 +259,5 @@ impl<TInvertedIndex: InvertedIndex> SparseVectorIndex<TInvertedIndex> {
             }
         }
         unique_record_ids.len()
-    }
-
-    fn get_query_cardinality(
-        &self,
-        filter: &Filter,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<CardinalityEstimation> {
-        let vector_storage = self.vector_storage.borrow();
-        let id_tracker = self.id_tracker.borrow();
-        let available_vector_count = vector_storage.available_vector_count();
-        let query_point_cardinality = self
-            .payload_index
-            .borrow()
-            .with_view(|v| v.estimate_cardinality(filter, hw_counter))?;
-        Ok(adjust_to_available_vectors(
-            query_point_cardinality,
-            available_vector_count,
-            id_tracker.available_point_count(),
-        ))
-    }
-
-    // Search using raw scorer
-    fn search_scored(
-        &self,
-        query_vector: &QueryVector,
-        filter: Option<&Filter>,
-        top: usize,
-        prefiltered_points: &mut Option<Vec<PointOffsetType>>,
-        vector_query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<ScoredPointOffset>> {
-        let vector_storage = self.vector_storage.borrow();
-        let id_tracker = self.id_tracker.borrow();
-        let deleted_point_bitslice = vector_query_context
-            .deleted_points()
-            .unwrap_or(id_tracker.deleted_point_bitslice());
-
-        let is_stopped = vector_query_context.is_stopped();
-
-        let searcher = BatchFilteredSearcher::new(
-            &[query_vector],
-            &vector_storage,
-            None,
-            None,
-            top,
-            deleted_point_bitslice,
-            vector_query_context.hardware_counter(),
-        )?;
-        let hw_counter = vector_query_context.hardware_counter();
-        let mut results = match filter {
-            Some(filter) => {
-                let filtered_points = match prefiltered_points {
-                    // `prefiltered_points` always contains visible points only so we don't need additional filtering here.
-                    Some(filtered_points) => filtered_points.iter().copied(),
-                    None => {
-                        let filtered_points = self.payload_index.borrow().with_view(|v| {
-                            v.query_points(
-                                filter,
-                                &hw_counter,
-                                &is_stopped,
-                                vector_query_context.deferred_internal_id(),
-                            )
-                        })?;
-                        *prefiltered_points = Some(filtered_points);
-                        prefiltered_points.as_ref().unwrap().iter().copied()
-                    }
-                };
-                searcher.peek_top_iter(filtered_points, &is_stopped)?
-            }
-            None => {
-                searcher.peek_top_all(&is_stopped, vector_query_context.deferred_internal_id())?
-            }
-        };
-        let res = results.pop().expect("single element results");
-        Ok(res)
-    }
-
-    pub fn search_plain(
-        &self,
-        sparse_vector: &SparseVector,
-        filter: &Filter,
-        top: usize,
-        prefiltered_points: &mut Option<Vec<PointOffsetType>>,
-        vector_query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<ScoredPointOffset>> {
-        let vector_storage = self.vector_storage.borrow();
-        let id_tracker = self.id_tracker.borrow();
-
-        let is_stopped = vector_query_context.is_stopped();
-
-        let deleted_point_bitslice = vector_query_context
-            .deleted_points()
-            .unwrap_or(id_tracker.deleted_point_bitslice());
-        let deleted_vectors = vector_storage.deleted_vector_bitslice();
-
-        let hw_counter = vector_query_context.hardware_counter();
-
-        let ids = match prefiltered_points {
-            // Deferred points get filtered in the `None` case and are added to `prefiltered_points`.
-            // In the `Some` case, we iterate over this set of points,
-            // so no additional filtering is required in that case.
-            Some(filtered_points) => filtered_points.iter(),
-            None => {
-                let filtered_points = self.payload_index.borrow().with_view(|v| {
-                    v.query_points(
-                        filter,
-                        &hw_counter,
-                        &is_stopped,
-                        vector_query_context.deferred_internal_id(),
-                    )
-                })?;
-                *prefiltered_points = Some(filtered_points);
-                prefiltered_points.as_ref().unwrap().iter()
-            }
-        }
-        .copied()
-        .filter(|&idx| check_deleted_condition(idx, deleted_vectors, deleted_point_bitslice))
-        .collect_vec();
-
-        let sparse_vector = self.indices_tracker.remap_vector(sparse_vector.clone());
-        let memory_handle = self.scores_memory_pool.get();
-        let mut hw_counter = vector_query_context.hardware_counter();
-        let is_index_on_disk = self.config.index_type.is_on_disk();
-        if is_index_on_disk {
-            hw_counter.set_vector_io_read_multiplier(1);
-        } else {
-            hw_counter.set_vector_io_read_multiplier(0);
-        }
-
-        let mut search_context = SearchContext::new(
-            sparse_vector,
-            top,
-            &self.inverted_index,
-            memory_handle,
-            &is_stopped,
-            &hw_counter,
-        );
-        let search_result = search_context.plain_search(&ids);
-        Ok(search_result)
-    }
-
-    // search using sparse vector inverted index
-    fn search_sparse(
-        &self,
-        sparse_vector: &SparseVector,
-        filter: Option<&Filter>,
-        top: usize,
-        vector_query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<ScoredPointOffset>> {
-        let vector_storage = self.vector_storage.borrow();
-        let id_tracker = self.id_tracker.borrow();
-        let deleted_point_bitslice = vector_query_context
-            .deleted_points()
-            .unwrap_or(id_tracker.deleted_point_bitslice());
-        let deleted_vectors = vector_storage.deleted_vector_bitslice();
-
-        let not_deleted_condition = |idx: PointOffsetType| -> bool {
-            check_deleted_condition(idx, deleted_vectors, deleted_point_bitslice)
-        };
-
-        let is_stopped = vector_query_context.is_stopped();
-
-        let sparse_vector = self.indices_tracker.remap_vector(sparse_vector.clone());
-        let memory_handle = self.scores_memory_pool.get();
-        let mut hw_counter = vector_query_context.hardware_counter();
-        let is_index_on_disk = self.config.index_type.is_on_disk();
-        if is_index_on_disk {
-            hw_counter.set_vector_io_read_multiplier(1);
-        } else {
-            hw_counter.set_vector_io_read_multiplier(0);
-        }
-
-        let mut search_context = SearchContext::new(
-            sparse_vector,
-            top,
-            &self.inverted_index,
-            memory_handle,
-            &is_stopped,
-            &hw_counter,
-        );
-
-        match filter {
-            Some(filter) => self.payload_index.borrow().with_view(|v| {
-                let filter_context = v.filter_context(filter, &hw_counter)?;
-                let matches_filter_condition = |idx: PointOffsetType| -> bool {
-                    not_deleted_condition(idx) && filter_context.check(idx)
-                };
-                Ok(search_context.search(&matches_filter_condition))
-            }),
-            None => Ok(search_context.search(&not_deleted_condition)),
-        }
-    }
-
-    fn search_nearest_query(
-        &self,
-        vector: &SparseVector,
-        filter: Option<&Filter>,
-        top: usize,
-        prefiltered_points: &mut Option<Vec<PointOffsetType>>,
-        vector_query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<ScoredPointOffset>> {
-        if vector.is_empty() {
-            return Ok(vec![]);
-        }
-
-        match filter {
-            Some(filter) => {
-                // if cardinality is small - use plain search
-                let query_cardinality =
-                    self.get_query_cardinality(filter, &vector_query_context.hardware_counter())?;
-                let threshold = self
-                    .config
-                    .full_scan_threshold
-                    .unwrap_or(DEFAULT_SPARSE_FULL_SCAN_THRESHOLD);
-                if query_cardinality.max < threshold {
-                    let _timer =
-                        ScopeDurationMeasurer::new(&self.searches_telemetry.small_cardinality);
-                    self.search_plain(
-                        vector,
-                        filter,
-                        top,
-                        prefiltered_points,
-                        vector_query_context,
-                    )
-                } else {
-                    let _timer =
-                        ScopeDurationMeasurer::new(&self.searches_telemetry.filtered_sparse);
-                    self.search_sparse(vector, Some(filter), top, vector_query_context)
-                }
-            }
-            None => {
-                let _timer = ScopeDurationMeasurer::new(&self.searches_telemetry.unfiltered_sparse);
-                self.search_sparse(vector, filter, top, vector_query_context)
-            }
-        }
-    }
-
-    pub fn search_query(
-        &self,
-        query_vector: &QueryVector,
-        filter: Option<&Filter>,
-        top: usize,
-        prefiltered_points: &mut Option<Vec<PointOffsetType>>,
-        vector_query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<ScoredPointOffset>> {
-        if top == 0 {
-            return Ok(vec![]);
-        }
-
-        match query_vector {
-            QueryVector::Nearest(vector) => self.search_nearest_query(
-                vector.try_into()?,
-                filter,
-                top,
-                prefiltered_points,
-                vector_query_context,
-            ),
-            QueryVector::RecommendBestScore(_)
-            | QueryVector::RecommendSumScores(_)
-            | QueryVector::Discover(_)
-            | QueryVector::Context(_)
-            | QueryVector::FeedbackNaive(_) => {
-                let _timer = if filter.is_some() {
-                    ScopeDurationMeasurer::new(&self.searches_telemetry.filtered_plain)
-                } else {
-                    ScopeDurationMeasurer::new(&self.searches_telemetry.unfiltered_plain)
-                };
-                self.search_scored(
-                    query_vector,
-                    filter,
-                    top,
-                    prefiltered_points,
-                    vector_query_context,
-                )
-            }
-        }
-    }
-}
-
-impl<TInvertedIndex: InvertedIndex> VectorIndexRead for SparseVectorIndex<TInvertedIndex> {
-    fn search(
-        &self,
-        vectors: &[&QueryVector],
-        filter: Option<&Filter>,
-        top: usize,
-        _params: Option<&SearchParams>,
-        query_context: &VectorQueryContext,
-    ) -> OperationResult<Vec<Vec<ScoredPointOffset>>> {
-        let mut results = Vec::with_capacity(vectors.len());
-        let mut prefiltered_points = None;
-
-        debug_assert_eq!(
-            self.deferred_internal_id,
-            query_context.deferred_internal_id(),
-            "SparseIndex and VectorQueryContext deferred_internal_id consistency violated."
-        );
-
-        for vector in vectors {
-            check_process_stopped(&query_context.is_stopped())?;
-
-            let search_results = if query_context.is_require_idf() {
-                let vector = (*vector).clone().transform(|mut vector| {
-                    match &mut vector {
-                        VectorInternal::Dense(_) | VectorInternal::MultiDense(_) => {
-                            return Err(OperationError::WrongSparse);
-                        }
-                        VectorInternal::Sparse(sparse) => {
-                            query_context.remap_idf_weights(&sparse.indices, &mut sparse.values)
-                        }
-                    }
-
-                    Ok(vector)
-                })?;
-
-                self.search_query(&vector, filter, top, &mut prefiltered_points, query_context)?
-            } else {
-                self.search_query(vector, filter, top, &mut prefiltered_points, query_context)?
-            };
-
-            results.push(search_results);
-        }
-        Ok(results)
-    }
-
-    fn get_telemetry_data(&self, detail: TelemetryDetail) -> VectorIndexSearchesTelemetry {
-        self.searches_telemetry.get_telemetry_data(detail)
-    }
-
-    fn indexed_vector_count(&self) -> usize {
-        self.inverted_index.vector_count()
-    }
-
-    fn size_of_searchable_vectors_in_bytes(&self) -> usize {
-        self.inverted_index.total_sparse_vectors_size()
-    }
-
-    /// Update statistics for idf-dot similarity.
-    fn fill_idf_statistics(
-        &self,
-        idf: &mut HashMap<DimId, usize>,
-        hw_counter: &HardwareCounterCell,
-    ) {
-        for (dim_id, count) in idf.iter_mut() {
-            if let Some(remapped_dim_id) = self.indices_tracker.remap_index(*dim_id)
-                && let Some(posting_list_len) = self
-                    .inverted_index
-                    .posting_list_len(&remapped_dim_id, hw_counter)
-            {
-                *count += posting_list_len
-            }
-        }
-    }
-
-    fn is_index(&self) -> bool {
-        true
-    }
-}
-
-impl<TInvertedIndex: InvertedIndex> VectorIndex for SparseVectorIndex<TInvertedIndex> {
-    fn files(&self) -> Vec<PathBuf> {
-        let config_file = SparseIndexConfig::get_config_path(&self.path);
-        if !config_file.exists() {
-            return vec![];
-        }
-
-        let mut all_files = vec![
-            IndicesTracker::file_path(&self.path),
-            self.path.join(VERSION_FILE),
-        ];
-        all_files.retain(|f| f.exists());
-
-        all_files.push(config_file);
-        all_files.extend_from_slice(&TInvertedIndex::files(&self.path));
-        all_files
-    }
-
-    fn immutable_files(&self) -> Vec<PathBuf> {
-        let config_file = SparseIndexConfig::get_config_path(&self.path);
-        if !config_file.exists() {
-            return vec![];
-        }
-
-        let mut immutable_files = vec![
-            self.path.join(VERSION_FILE), // TODO: Is version file immutable?
-        ];
-        immutable_files.retain(|f| f.exists());
-
-        immutable_files.push(config_file);
-        immutable_files.extend_from_slice(&TInvertedIndex::immutable_files(&self.path));
-        immutable_files
-    }
-
-    fn update_vector(
-        &mut self,
-        id: PointOffsetType,
-        vector: Option<VectorRef>,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        let (old_vector, new_vector) = {
-            let mut vector_storage = self.vector_storage.borrow_mut();
-            let old_vector = vector_storage
-                .get_vector_opt::<Random>(id)
-                .map(CowVector::to_owned);
-            let new_vector = if let Some(vector) = vector {
-                vector_storage.insert_vector(id, vector, hw_counter)?;
-                vector.to_owned()
-            } else {
-                let default_vector = vector_storage.default_vector();
-                if id as usize >= vector_storage.total_vector_count() {
-                    // Vector doesn't exist in the storage
-                    // Insert default vector to keep the sequence
-                    vector_storage.insert_vector(
-                        id,
-                        VectorRef::from(&default_vector),
-                        hw_counter,
-                    )?;
-                }
-                vector_storage.delete_vector(id)?;
-                default_vector
-            };
-            (old_vector, new_vector)
-        };
-
-        if self.config.index_type != SparseIndexType::MutableRam {
-            return Err(OperationError::service_error(
-                "Cannot update vector in non-appendable index",
-            ));
-        }
-
-        let point_is_deferred = self
-            .deferred_internal_id
-            .is_some_and(|deferred| id >= deferred);
-
-        if point_is_deferred {
-            return Ok(());
-        }
-
-        let vector = SparseVector::try_from(new_vector)?;
-        let old_vector: Option<SparseVector> =
-            old_vector.map(SparseVector::try_from).transpose()?;
-
-        // do not upsert empty or deferred vectors into the index
-        if !vector.is_empty() {
-            self.indices_tracker.register_indices(&vector);
-            let vector = self.indices_tracker.remap_vector(vector);
-            let old_vector = old_vector.map(|v| self.indices_tracker.remap_vector(v));
-            self.inverted_index.upsert(id, vector, old_vector);
-        } else if let Some(old_vector) = old_vector {
-            // Make sure empty vectors do not interfere with the index
-            if !old_vector.is_empty() {
-                let old_vector = self.indices_tracker.remap_vector(old_vector);
-                self.inverted_index.remove(id, old_vector);
-            }
-        }
-
-        Ok(())
     }
 }
